requirements.txt
pycryptodome>=3.19.1
pillow>=10.2.0
gmssl>=3.2.1
gmpy2>=2.1.5 